requirements.txt
appdirs==1.4.4
argparse==1.4.0
attrs==22.1.0 ; python_version >= '3.5'
beautifulsoup4==4.11.1
censys==2.1.9 ; python_version > '3.6'
IPy
minio
jinja2
communityid
dnspython==2.2.1
mattermostdriver==7.3.2
nfstream
python-magic==0.4.27
python-whois==0.8.0
redis==4.3.4 ; python_version >= '3.6'
requests
rich==12.5.1 ; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
tornado==6.2 ; python_version >= '3.7'
tzdata==2022.4 ; python_version >= '3.6'
tzlocal==4.2 ; python_version >= '3.6'
yara-python==3.8.1